+++ src/rpc/methods/f3/types.rs
@@ -196,7 +196,11 @@ pub struct GpbftConfig {
     #[serde(with = "crate::lotus_json")]
     pub delta: Duration,
     pub delta_back_off_exponent: f64,
+    pub quality_delta_multiplier: f64,
     pub max_lookahead_rounds: u64,
+
+    pub chain_proposed_length: usize,
+
     #[schemars(with = "u64")]
     #[serde(with = "crate::lotus_json")]
     pub rebroadcast_backoff_base: Duration,
@@ -237,6 +241,28 @@ pub struct CertificateExchangeConfig {
     pub maximum_poll_interval: Duration,
 }
 
+#[derive(PartialEq, Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[serde(rename_all = "PascalCase")]
+pub struct PubSubConfig {
+    pub compression_enabled: bool,
+}
+
+#[derive(PartialEq, Debug, Clone, Serialize, Deserialize, JsonSchema)]
+#[serde(rename_all = "PascalCase")]
+pub struct ChainExchangeConfig {
+    pub subscription_buffer_size: usize,
+    pub max_chain_length: usize,
+    pub max_instance_lookahead: usize,
+    pub max_discovered_chains_per_instance: usize,
+    pub max_wanted_chains_per_instance: usize,
+    #[schemars(with = "u64")]
+    #[serde(with = "crate::lotus_json")]
+    pub rebroadcast_interval: Duration,
+    #[schemars(with = "u64")]
+    #[serde(with = "crate::lotus_json")]
+    pub max_timestamp_age: Duration,
+}
+
 #[serde_as]
 #[derive(PartialEq, Debug, Clone, Serialize, Deserialize, JsonSchema)]
 #[serde(rename_all = "PascalCase")]
@@ -262,6 +288,8 @@ pub struct F3Manifest {
     #[serde(rename = "EC")]
     pub ec: EcConfig,
     pub certificate_exchange: CertificateExchangeConfig,
+    pub pub_sub: PubSubConfig,
+    pub chain_exchange: ChainExchangeConfig,
 }
 lotus_json_with_self!(F3Manifest);
 
@@ -677,7 +705,7 @@ mod tests {
         // lotus f3 manifest --output json
         let lotus_json = serde_json::json!({
           "Pause": false,
-          "ProtocolVersion": 4,
+          "ProtocolVersion": 5,
           "InitialInstance": 0,
           "BootstrapEpoch": 2081674,
           "NetworkName": "calibrationnet",
@@ -691,7 +719,9 @@ mod tests {
           "Gpbft": {
             "Delta": 6000000000_u64,
             "DeltaBackOffExponent": 2_f64,
+            "QualityDeltaMultiplier": 1_f64,
             "MaxLookaheadRounds": 5,
+            "ChainProposedLength": 100,
             "RebroadcastBackoffBase": 6000000000_u64,
             "RebroadcastBackoffExponent": 1.3,
             "RebroadcastBackoffSpread": 0.1,
@@ -711,7 +741,7 @@ mod tests {
               6.27,
               7.5
             ],
-            "HeadLookback": 0,
+            "HeadLookback": 4,
             "Finalize": true
           },
           "CertificateExchange": {
@@ -719,6 +749,18 @@ mod tests {
             "ServerRequestTimeout": 60000000000_u64,
             "MinimumPollInterval": 30000000000_u64,
             "MaximumPollInterval": 120000000000_u64
+          },
+          "PubSub": {
+            "CompressionEnabled": false
+          },
+          "ChainExchange": {
+            "SubscriptionBufferSize": 32,
+            "MaxChainLength": 100,
+            "MaxInstanceLookahead": 10,
+            "MaxDiscoveredChainsPerInstance": 1000,
+            "MaxWantedChainsPerInstance": 1000,
+            "RebroadcastInterval": 2000000000_u64,
+            "MaxTimestampAge": 8000000000_u64
           }
         });
         let manifest: F3Manifest = serde_json::from_value(lotus_json.clone()).unwrap();